# unify compare kernels