--- a/aten/src/ATen/native/cuda/CompareGEKernel.cu
+++ b/aten/src/ATen/native/cuda/CompareGEKernel.cu
@@ -10,19 +10,22 @@
 // of a __device__ lambda not have internal linkage.
 
 namespace at { namespace native { namespace {
+
+enum class OpType {GE, GT, LE, LT};
+
 template<typename scalar_t>
 struct CompareFunctor{
-  CompareFunctor(const int op): op_(op) {TORCH_INTERNAL_ASSERT_DEBUG_ONLY(op_>=0 && op_ <= 3);}
-  const int op_;
+  CompareFunctor(OpType op): op_(op) {};
+  OpType op_;
   __device__ __forceinline__ bool operator() (scalar_t a, scalar_t b) const {
     //printf("vals %ld %ld\n", a, b);
-    if (op_ == 0) {
+    if (op_ == OpType::GE) {
       return a >= b;
-    } else if (op_ == 1) {
+    } else if (op_ == OpType::GT) {
       return a > b;
-    } else if (op_ == 2) {
+    } else if (op_ == OpType::LE) {
       return a <= b;
-    } else if (op_ == 3) { //LT
+    } else { //LT
       return a < b;
     }
   }
@@ -32,25 +35,25 @@ struct CompareFunctor{
 
 void ge_kernel_cuda(TensorIteratorBase& iter) {
   AT_DISPATCH_ALL_TYPES_AND3(kHalf, kBFloat16, kBool, iter.common_dtype(), "ge_cuda", [&]() {
-    gpu_kernel_with_scalars(iter, CompareFunctor<scalar_t>(0));
+    gpu_kernel_with_scalars(iter, CompareFunctor<scalar_t>(OpType::GE));
   });
 }
 
 void gt_kernel_cuda(TensorIteratorBase& iter) {
   AT_DISPATCH_ALL_TYPES_AND3(kHalf, kBFloat16, kBool, iter.common_dtype(), "ge_cuda", [&]() {
-    gpu_kernel_with_scalars(iter, CompareFunctor<scalar_t>(1));
+    gpu_kernel_with_scalars(iter, CompareFunctor<scalar_t>(OpType::GT));
   });
 }
 
 void le_kernel_cuda(TensorIteratorBase& iter) {
   AT_DISPATCH_ALL_TYPES_AND3(kHalf, kBFloat16, kBool, iter.common_dtype(), "ge_cuda", [&]() {
-    gpu_kernel_with_scalars(iter, CompareFunctor<scalar_t>(2));
+    gpu_kernel_with_scalars(iter, CompareFunctor<scalar_t>(OpType::LE));
   });
 }
 
 void lt_kernel_cuda(TensorIteratorBase& iter) {
   AT_DISPATCH_ALL_TYPES_AND3(kHalf, kBFloat16, kBool, iter.common_dtype(), "ge_cuda", [&]() {
-    gpu_kernel_with_scalars(iter, CompareFunctor<scalar_t>(3));
+    gpu_kernel_with_scalars(iter, CompareFunctor<scalar_t>(OpType::LT));
   });
 }
 
@@ -58,21 +61,5 @@ REGISTER_DISPATCH(ge_stub, &ge_kernel_cuda);
 REGISTER_DISPATCH(gt_stub, &gt_kernel_cuda);
 REGISTER_DISPATCH(le_stub, &le_kernel_cuda);
 REGISTER_DISPATCH(lt_stub, &lt_kernel_cuda);
-
-
-// template<typename scalar_t>
-// struct CompareGEFunctor {
-//   __device__ __forceinline__ bool operator() (scalar_t a, scalar_t b) const {
-//     return a >= b;
-//   }
-// };
-
-// void ge_kernel_cuda(TensorIteratorBase& iter) {
-//   AT_DISPATCH_ALL_TYPES_AND3(kHalf, kBFloat16, kBool, iter.common_dtype(), "ge_cuda", [&]() {
-//     gpu_kernel_with_scalars(iter, CompareGEFunctor<scalar_t>());
-//   });
-// }
-
-// REGISTER_DISPATCH(ge_stub, &ge_kernel_cuda);
 
 }} // namespace at::native