--- a/aten/src/ATen/native/cuda/CompareLTKernel.cu
+++ b/aten/src/ATen/native/cuda/CompareLTKernel.cu
@@ -11,19 +11,19 @@
 
 namespace at { namespace native {
 
-template<typename scalar_t>
-struct CompareLTFunctor {
-  __device__ __forceinline__ bool operator() (scalar_t a, scalar_t b) const {
-    return a < b;
-  }
-};
+// template<typename scalar_t>
+// struct CompareLTFunctor {
+//   __device__ __forceinline__ bool operator() (scalar_t a, scalar_t b) const {
+//     return a < b;
+//   }
+// };
 
-void lt_kernel_cuda(TensorIteratorBase& iter) {
-  AT_DISPATCH_ALL_TYPES_AND3(kHalf, kBFloat16, kBool, iter.common_dtype(), "lt_cuda", [&]() {
-    gpu_kernel_with_scalars(iter, CompareLTFunctor<scalar_t>());
-  });
-}
+// void lt_kernel_cuda(TensorIteratorBase& iter) {
+//   AT_DISPATCH_ALL_TYPES_AND3(kHalf, kBFloat16, kBool, iter.common_dtype(), "lt_cuda", [&]() {
+//     gpu_kernel_with_scalars(iter, CompareLTFunctor<scalar_t>());
+//   });
+// }
 
-REGISTER_DISPATCH(lt_stub, &lt_kernel_cuda);
+// REGISTER_DISPATCH(lt_stub, &lt_kernel_cuda);
 
 }} // namespace at::native